projectV2/requirements.txt
numpy>=1.24.0
opencv-python>=4.8.0
insightface>=0.7.3
onnxruntime>=1.16.0
scikit-learn>=1.3.0
scikit-image>=0.21.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
pyyaml>=6.0
joblib>=1.3.0
scipy>=1.11.0
Pillow>=10.0.0
